Massive core and plugin update

+++ php/wp-content/plugins/wp-fastest-cache/wpFastestCache.php
@@ -3,7 +3,7 @@
 Plugin Name: WP Fastest Cache
 Plugin URI: http://wordpress.org/plugins/wp-fastest-cache/
 Description: The simplest and fastest WP Cache system
-Version: 0.8.5.1
+Version: 0.8.5.4
 Author: Emre Vona
 Author URI: http://tr.linkedin.com/in/emrevona
 
@@ -97,33 +97,21 @@ GNU General Public License for more details.
 						exit;
 					}
 				}
-			}else if(isset($_GET) && isset($_GET["action"]) && $_GET["action"] == "wpfc_download_premium"){
-				$res = array();
-				$response = wp_remote_get("http://api.wpfastestcache.net/premium/download/".str_replace(array("http://", "www."), "", $_SERVER["HTTP_HOST"])."/".get_option("WpFc_api_key"), array('timeout' => 10 ) );
+			}else if(isset($_GET) && isset($_GET["action"]) && $_GET["action"] == "wpfc_update_premium"){
+				if($this->isPluginActive("wp-fastest-cache-premium/wpFastestCachePremium.php")){
+					if(!file_exists(WPFC_WP_PLUGIN_DIR."/wp-fastest-cache-premium/pro/library/update.php")){
+						$res = array("success" => false, "error_message" => "update.php is not exist");
+					}else{
+						include_once $this->get_premium_path("update.php");
+						
+						if(!class_exists("WpFastestCacheUpdate")){
+							$res = array("success" => false, "error_message" => "WpFastestCacheUpdate is not exist");
+						}else{
+							$wpfc_premium = new WpFastestCacheUpdate();
+							$content = $wpfc_premium->download_premium();
 
-				if ( !$response || is_wp_error( $response ) ) {
-					$res = array("success" => false, "error_message" => $response->get_error_message());
-				}else{
-					if(wp_remote_retrieve_response_code($response) == 200){
-						$wpfc_premium_download_link = wp_remote_retrieve_body( $response );
-						if($wpfc_premium_download_link){
-
-							$response_download = wp_remote_get($wpfc_premium_download_link);
-
-							if ( !$response_download || is_wp_error( $response_download ) ) {
-								$res = array("success" => false, "error_message" => $response_download->get_error_message());
-							}else{
-								if(wp_remote_retrieve_response_code($response_download) == 200){
-									if(!$wpfc_zip_data = wp_remote_retrieve_body( $response_download )){
-										$res = array("success" => false, "error_message" => ".zip file is empty");
-									}
-								}else{
-									$res = array("success" => false, "error_message" => "Download Source is unavailable");
-								}
-							}
-
-
-							if($wpfc_zip_data){
+							if($content["success"]){
+								$wpfc_zip_data = $content["content"];
 
 								$wpfc_zip_dest_path = WPFC_WP_PLUGIN_DIR."/wp-fastest-cache-premium.zip";
 
@@ -135,38 +123,42 @@ GNU General Public License for more details.
 									if(function_exists("unzip_file")){
 										$this->rm_folder_recursively(WPFC_WP_PLUGIN_DIR."/wp-fastest-cache-premium");
 										
-										WP_Filesystem();
-										$unzipfile = unzip_file($wpfc_zip_dest_path, WPFC_WP_PLUGIN_DIR."/");
+										if(!function_exists('gzopen')){
+											$res = array("success" => false, "error_message" => "Missing zlib extension"); 
+										}else{
+											WP_Filesystem();
+											$unzipfile = unzip_file($wpfc_zip_dest_path, WPFC_WP_PLUGIN_DIR."/");
 
+											if ($unzipfile) {
+												$result = activate_plugin( 'wp-fastest-cache-premium/wpFastestCachePremium.php' );
 
-										if ($unzipfile) {
-											$result = activate_plugin( 'wp-fastest-cache-premium/wpFastestCachePremium.php' );
-
-											if ( is_wp_error( $result ) ) {
-												$res = array("success" => false, "error_message" => "Error occured while the plugin was activated", "error_code" => 1, "file_url" => $wpfc_premium_download_link); 
-											}else{
-												$res = array("success" => true);
-												$this->deleteCache(true);
+												if ( is_wp_error( $result ) ) {
+													$res = array("success" => false, "error_message" => "Error occured while the plugin was activated"); 
+												}else{
+													$res = array("success" => true);
+													$this->deleteCache(true);
+												}
+											} else {
+												$res = array("success" => false, "error_message" => 'Error occured while the file was unzipped');      
 											}
-										} else {
-											$res = array("success" => false, "error_message" => 'Error occured while the file was unzipped', "error_code" => 2, "file_url" => $wpfc_premium_download_link);      
 										}
+										
 									}else{
-										$res = array("success" => false, "error_message" => "unzip_file() is not found", "error_code" => 3, "file_url" => $wpfc_premium_download_link);
+										$res = array("success" => false, "error_message" => "unzip_file() is not found");
 									}
 								}else{
-									$res = array("success" => false, "error_message" => "/wp-content/plugins/ is not writable", "error_code" => 4, "file_url" => $wpfc_premium_download_link);
+									$res = array("success" => false, "error_message" => "/wp-content/plugins/ is not writable");
 								}
 							}else{
-								$res = array("success" => false, "error_message" => "Error: Service is unavailable. Try later...");
+								$res = array("success" => false, "error_message" => $content["error_message"]);
 							}
-						}else{
-							$res = array("success" => false, "error_message" => "Error: Link is empty");
 						}
-					}else{
-						$res = array("success" => false, "error_message" => "Error: Try later...");
 					}
+				}else{
+					$res = array("success" => false, "error_message" => "Premium is not active");
+
 				}
+
 				echo json_encode($res);
 				exit;
 			}else if(isset($_GET) && isset($_GET["action"]) && in_array($_GET["action"], $optimize_image_ajax_requests)){
@@ -315,28 +307,32 @@ GNU General Public License for more details.
 		public function singleDeleteCache($comment_id = false, $post_id = false){
 			if($comment_id){
 				$comment = get_comment($comment_id);
+				
+				if($comment && $comment->comment_post_ID){
+					$post_id = $comment->comment_post_ID;
+				}
 			}
 
-			$post_id = $post_id ? $post_id : $comment->comment_post_ID;
+			if($post_id){
+				$permalink = get_permalink($post_id);
 
-			$permalink = get_permalink($post_id);
+				if(preg_match("/http:\/\/[^\/]+\/(.+)/", $permalink, $out)){
+					$path = $this->getWpContentDir()."/cache/all/".$out[1];
+					$mobile_path = $this->getWpContentDir()."/cache/wpfc-mobile-cache/".$out[1];
 
-			if(preg_match("/http:\/\/[^\/]+\/(.+)/", $permalink, $out)){
-				$path = $this->getWpContentDir()."/cache/all/".$out[1];
-				$mobile_path = $this->getWpContentDir()."/cache/wpfc-mobile-cache/".$out[1];
+					if(is_dir($path)){
+						if($this->isPluginActive("wp-fastest-cache-premium/wpFastestCachePremium.php")){
+							include_once $this->get_premium_path("logs.php");
+							$log = new WpFastestCacheLogs("delete");
+							$log->action();
+						}
 
-				if(is_dir($path)){
-					if($this->isPluginActive("wp-fastest-cache-premium/wpFastestCachePremium.php")){
-						include_once $this->get_premium_path("logs.php");
-						$log = new WpFastestCacheLogs("delete");
-						$log->action();
+						$this->rm_folder_recursively($path);
 					}
 
-					$this->rm_folder_recursively($path);
-				}
-
-				if(is_dir($mobile_path)){
-					$this->rm_folder_recursively($mobile_path);
+					if(is_dir($mobile_path)){
+						$this->rm_folder_recursively($mobile_path);
+					}
 				}
 			}
 		}
@@ -374,7 +370,7 @@ GNU General Public License for more details.
 
 			if(is_dir($minified_cache_path)){
 				if($minified){
-					rename($minified_cache_path, $this->getWpContentDir()."/cache/tmpWpfc/".time());
+					rename($minified_cache_path, $this->getWpContentDir()."/cache/tmpWpfc/m".time());
 					$deleted = true;
 				}
 			}
@@ -444,7 +440,7 @@ GNU General Public License for more details.
 		    return $path;
 		}
 
-		protected function rm_folder_recursively($dir, $i = 1) {
+		public function rm_folder_recursively($dir, $i = 1) {
 			$files = @scandir($dir);
 		    foreach((array)$files as $file) {
 		    	if($i > 500){
